Reorganization + bugfix of template/page editing code.
