+++ commands/miscinfo/action/edit_ids_on_page.py
@@ -1,18 +1,18 @@
 from commands.miscinfo.utils import track_page_miscinfo as track_page
+from commands.miscinfo.utils import track_page_edit as track_edit
 from common_utils.file_reader import read_csv_file
 from tockdomio import tockdomread, tockdomwrite
 
-import warnings
-
-def add_ids_by_pageid(page_id, page_text, new_arguments: dict, update_existing=False):
+def add_ids_by_pageid(page_id, page_text, new_arguments: dict, update_wbz=False):
+    print(new_arguments)
     arguments = track_page.get_miscinfo_template(page_text)
-    track_page.patch_miscinfo_template(arguments, new_arguments, update_existing)
-    template_text = track_page.create_miscinfo_template(arguments)
+    track_edit.patch_ids_to_miscinfo_template(arguments, new_arguments, update_wbz)
+    template_text = track_edit.create_miscinfo_template(arguments)
 
-    section_text = str(track_page.replace_miscinfo_template(page_text, template_text))
+    section_text = str(track_edit.replace_miscinfo_template(page_text, template_text))
     print(section_text)
 
-    response = tockdomwrite.edit_section(page_id, 0, section_text, "Add miscinfo template (via API)")
+    response = tockdomwrite.edit_section(page_id, 0, section_text, "Add wbz and image ids to template (via API)")
     print(response.json())
     was_successful = response.json()["edit"]["result"] == "Success"
     return was_successful
