+++ commands/miscinfo/utils/track_page_miscinfo.py
@@ -1,4 +1,3 @@
-from mediawiki.mediawiki_create import *
 from mediawiki.mediawiki_parse import *
 import warnings
 
@@ -24,31 +23,3 @@ def get_miscinfo_template(page_text):
         arguments[argument.name] = argument.value.strip()
 
     return arguments
-
-def patch_miscinfo_template(arguments, new_arguments, update_existing):
-    for new_arg, new_val in new_arguments.items():
-        if new_arg not in arguments:
-            warnings.warn("{} not a valid argument!".format(new_arg))
-            continue
-
-        would_replace_existing = (new_arg in arguments) and arguments[new_arg] and arguments[new_arg] != new_val
-        if update_existing and would_replace_existing:
-            warnings.warn("{} already exists and update flag is false".format(new_arg))
-            continue
-
-        if new_val:
-            arguments[new_arg] = new_val
-
-def create_miscinfo_template(arguments: dict):
-    return create_template_from_args(arguments, "Misc-Info")
-
-def replace_miscinfo_template(page_text, template_text):
-    section = get_first_section_from_page(page_text)
-    template: Template = get_template_with_name(section, "Misc-Info")
-
-    if len(section.contents) == len(template) + 2:
-        section.contents = template_text + "\n"
-    else:
-        template.string = template_text
-
-    return str(section)
